Former-commit-id: 16df1056ca157598e6dfc19a16e2a9a055016dea [formerly 0242380c7647a204b8eb9e5890b4323825e88dbb]
Former-commit-id: 6e813ab884a3fb4fd7024df21989036b44a798d7

+++ bridge/bindings/jsc/DOM/event_target.h
@@ -76,7 +76,7 @@ public:
     bool dispatchEvent(EventInstance *event);
 
     ~EventTargetInstance() override;
-    int64_t eventTargetId;
+    int32_t eventTargetId;
     NativeEventTarget *nativeEventTarget {nullptr};
 
   private:
